pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["aasherkamal216 <aasherkamal786@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
